Task #69 - main screen (#99)
* Issue #67 First version of main screen

screen xml and i18n related to it.
Simply added buttons with labels to the screen.

Still missing navigation or any logic. There is no layout as well

* Adapting to android naming guide

Changes to name of resources, extraction of i18n strings
resolves #67

* Clean up of activities and navigation

new activity showed up inn the mean time, for task creation. This commit
creates a new activity class, clears up navigation between activities

* fixing failing test

fixes #69
