+++ Friendly-plans/app/src/main/AndroidManifest.xml
@@ -13,10 +13,11 @@
       android:screenOrientation="landscape">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
-
-        <category android:name="android.intent.category.LAUNCHER"/>
-      </intent-filter>
-    </activity>
-  </application>
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity android:name=".TaskCreateActivity">
+        </activity>
+    </application>
 
 </manifest>
